Adjust node name adjustments for vanilla fix
