+++ CommunityBugFixCollection/CommunityBugFixCollection.csproj
@@ -30,11 +30,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="MicroUtils.HarmonyAnalyzers" Version="1.4.0">
+    <PackageReference Include="MicroUtils.HarmonyAnalyzers" Version="1.6.0-beta1744347465">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="MonkeyLoader.GamePacks.Resonite" Version="0.22.1-beta" />
+    <PackageReference Include="MonkeyLoader.GamePacks.Resonite" Version="0.23.1-beta" />
     <PackageReference Include="PolySharp" Version="1.15.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -42,11 +42,11 @@
     <PackageReference Include="Resonite.Elements.Assets" Version="1.3.3" />
     <PackageReference Include="Resonite.Elements.Core" Version="1.4.3" />
     <PackageReference Include="Resonite.Elements.Quantity" Version="1.2.3" />
-    <PackageReference Include="Resonite.FrooxEngine" Version="2025.5.14.23" />
+    <PackageReference Include="Resonite.FrooxEngine" Version="2025.5.33.1285" />
     <PackageReference Include="Resonite.FrooxEngine.Store" Version="1.0.5" />
     <PackageReference Include="Resonite.ProtoFlux.Core" Version="1.3.1" />
     <PackageReference Include="Resonite.ProtoFlux.Nodes.Core" Version="1.3.1" />
-    <PackageReference Include="Resonite.ProtoFluxBindings" Version="2025.5.18.24" />
+    <PackageReference Include="Resonite.ProtoFluxBindings" Version="2025.5.29.1285" />
     <PackageReference Include="Resonite.SkyFrost.Base" Version="2.1.0" />
     <PackageReference Include="Resonite.SkyFrost.Base.Models" Version="2.1.5" />
   </ItemGroup>
