+++ CommunityBugFixCollection/NodeNameAdjustments.cs
@@ -1,6 +1,5 @@
 ﻿using FrooxEngine.ProtoFlux;
 using FrooxEngine.ProtoFlux.Runtimes.Execution.Nodes.Math;
-using FrooxEngine.ProtoFlux.Runtimes.Execution.Nodes.Strings.Characters;
 using HarmonyLib;
 using System;
 using System.Collections.Generic;
@@ -13,48 +12,6 @@ namespace CommunityBugFixCollection
     internal sealed class NodeNameAdjustments : ResoniteBugFixMonkey<NodeNameAdjustments>
     {
         public override IEnumerable<string> Authors => Contributors.Banane9;
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(Backspace))]
-        private static string BackspaceNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Backspace (\\b)";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(Bell))]
-        private static string BellNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Bell (\\a)";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(CarriageReturn))]
-        private static string CarriageReturnNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Carriage Return (\\r)";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(FormFeed))]
-        private static string FormFeedNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Form Feed (\\f)";
-        }
-
-        private static bool Prepare() => Enabled;
 
         [HarmonyPostfix]
         [HarmonyPatch(typeof(Remap11_01_Double))]
@@ -74,36 +31,6 @@ namespace CommunityBugFixCollection
                 return __result;
 
             return "Remap [-1; 1] to [0; 1]";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(Space))]
-        private static string SpaceNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Space ( )";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(Tab))]
-        private static string TabNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Tab (\\t)";
-        }
-
-        [HarmonyPostfix]
-        [HarmonyPatch(typeof(VerticalTab))]
-        private static string VerticalTabNamePostfix(string __result)
-        {
-            if (!Enabled)
-                return __result;
-
-            return "Vertical Tab (\\v)";
         }
     }
 }
